modelExecutorLambda/requirements.txt
boto3
httpx
pandas
google-generativeai